Documentation: Fix broken links
This change mostly changes  links that were identified as broken by
the 'website_scans' jenkins job.

There were some links that seem to be up at times, but that are
identified by link-checker as broken because of SSL issues.

At least one other link was changed to point to archive.org so
that it doesn't break at some point in the future.  We should
probably try to make sure that everything is archived there and
point to those versions when possible.

There are still lots more links to do.

Signed-off-by: Martin Roth <martin@coreboot.org>
Change-Id: I36868ddf6113e18fa6841427dd635c75445b7bef
Reviewed-on: https://review.coreboot.org/c/coreboot/+/62672
Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Reviewed-by: Felix Singer <felixsinger@posteo.net>